Get JavaScript datatable containing bids and asks for all 'PredictIt' markets

Wrapper function to return a datatable containing 'PredictIt' market data. Can be displayed in a shinyApp, RMarkdown document, or exported via saveWidget().

markets_table(data)

Arguments

data

'PredictIt' market data, of class data.frame or tibble, as returned by all_markets() or single_market.

Value

An interactive datatable object containing formatted bid and ask data for the provided market data.

Examples

data <- all_markets()
#> Warning: Error fetching PredictIt market data 
#> Check your internet connection and try again 
#> 
#> Note: do not abuse the API by requesting data more than once every 60 seconds
markets_table(data)